Allow arguments to be passed to Docker image (#301)
* docker: EntryPoint instead of Cmd

The value of CMD gets overwritten by arguments passed to `docker run`,
which is not what we want.

* docker: use contents

Makes centrifuge-chain available under /bin

* flake: use legacyPackages

Internally, it does an import, see flake.nix in nixpkgs.

* ci: add hours, seconds to docker tag

* ci: only run docker steps on upstream

This prevents steps that would certainly fail from running on forked repos.

* docker: add centrifuge-chain to PATH

* nix: avoid rebuilding when source hasn't changed

This is achieved by setting src to ./., rather than self. The previous
approach meant that changes to the flake were seen as changes to the
package's src, and thus triggered a rebuild.

* git: ignore nix's result symlink

* nix: formatting

* nix: simplify outputs function's arguments

* nix: avoid over-using with

It was bringing all of nixpkgs' attrset in scope, which can create
hard-to-debug clashes.

* nix: add explitit package to flake

So that it's more easily discoverable.

* ci: move docker step to docker job

* nix: fix PATH in Docker image

* nix: run centrifuge-chain from PATH in Docker img

* Revert "nix: avoid rebuilding when source hasn't changed"

This reverts commit 9c09de7c24d89d22fc6e2d7983c275cafd338f22.

This was not actually sufficient, since we need to also "ignore" changes
to flake.nix (which cannot be gitignored) - otherwise changes to it will
still trigger a rebuild of the rust package.

Will handle this in a separate branch.