ci: run on direct pushes, but only to master
This was otherwise making CI runs for PRs twice as long, as they were
being built twice (once for the PR event, once for the push event).

We instead assume that branches will always be accompanied by a PR,
whereas hotfixes to master will be checked on push as well.