Adding conversionRate native-to-relay above 1
The pallet is now able to convert with a ratio above and below one.
In order so allow this functionality an additional associated type
is intrioduced, that provideds the capability to convert a u64
into the native balance type of the chain. This is needed as our
18 deciamls systems is to large to present a reasonable conversion
rate with a u32.

Additonally the tests have been adjusted to the changes.

Furthermore, in order to prevent usage of the GenesisConfig, the
current code uses storageValues with reasonable default values.

NOTE: These defaults are probably going to change before the deploy.