Integrated child tree proof (#369)
* Integrated child tree proof

The following commit integrates the verification of the
contribution via the provided child tree proof. Hence, the reward
interface is adjusted, so that the right proof is provided.

Furthermore the storage is adjusted to hold the trie index, while
the participant registration part is removedIntegrated child tree proof

* Removed tick in favor of proof validations

With the current implementation we favor proof validation upon
transaction validation instead of providing a window for submitting
transactions. This puts the workload to the validators instead of
malicious behaviour.

* [WIP] Intermediate Commit for Testing

* Integrate PR recommendations