Feature/crowdloan/reward (#342)
* Adding conversionRate native-to-relay above 1

The pallet is now able to convert with a ratio above and below one.
In order so allow this functionality an additional associated type
is intrioduced, that provideds the capability to convert a u64
into the native balance type of the chain. This is needed as our
18 deciamls systems is to large to present a reasonable conversion
rate with a u32.

Additonally the tests have been adjusted to the changes.

Furthermore, in order to prevent usage of the GenesisConfig, the
current code uses storageValues with reasonable default values.

NOTE: These defaults are probably going to change before the deploy.

* Catches error upon failure of vested_transfer

Description is in the code. But as an essence, we must ensure that
both extriniscs are executed successfully. Hence, we must first
transfer and then vest transfer, as the former is revertable.

* Added check before execution if vested_transfer can fail

Adjusted the tests accordingly. No, in case of failure, the reward
call is a noop.

* Fixing typos and review minors

* Reformat Code

* Added zero-direct test, Updated event names

* Rust format

* After rebase adjustments