Do not allow other proxy types to create NonProxy proxies
A NonProxy proxy has access to extrinsics which the other types
(besides Any) do not have access to. Therefore, other proxy types
which have access to proxy::addProxy should not be allowed to create
a NonProxy proxy. For now, this only applies to the NonTransfer
proxy type.

The inverse filtering is not needed, since NonProxy by definition
cannot create new proxies.