Use `ensure_can_withdraw` in chain bridge transfers
Previously, the chain bridge transfer checked that the total transfer
amount + fee was greater than the "free" balance in the account. This
does not take into account various types of locks that can be on an
account.

This PR changes to use the `ensure_can_withdraw` function from the
Currency trait. This function checks all locks and fails the
transaction if there is not enough available balance in the account.