Replace Vesting proxy with NonProxy proxy (#451)
* Adds the NonProxy Type to Proxies

This commit removes the Vesting type in favor of the NonProxy type.
The new proxy type forbids to an account of this type to execute
any proxy pallet related operations, besides executing as a different
call as a proxy himself.

* Simplify NonProxy filter expression

* Do not allow other proxy types to create NonProxy proxies

A NonProxy proxy has access to extrinsics which the other types
(besides Any) do not have access to. Therefore, other proxy types
which have access to proxy::addProxy should not be allowed to create
a NonProxy proxy. For now, this only applies to the NonTransfer
proxy type.

The inverse filtering is not needed, since NonProxy by definition
cannot create new proxies.

* Bump runtime version

Co-authored-by: mustermeiszer <mustermeiszer@posteo.de>