Migration pallet (#403)
* Initially setup for migration-manager pallet

This commit already contains the first try of migrating the system
account storage via an extrinsic.

* Fixed tests and pallet integration

* Add transactional, increase MaxAccount

* Clean up and vesting support

* Adding migration for vesting and proxies

Furthermore, due to the need for the proxy migration to have the
complete "ProxyType" enum available in the runtime,
this commit adds the "NonProxy" variant.

* Currenlty, final impl of migration.

The current impl contains, issuance, accounts, vesting and proxies.
In case, we are going to add more storage values to the migration
this needs to be extended.

Missing:
- tests

* Finished tests.

This commit contains the test data. Fetched from centrifuge mainnet
at the block of height 6400000.
Furthermore, the tests are implemented.

* Final clean-up after rebase

* PR request changes

* Benchmarking + benched weights

This commit adds the benchmarking section to the pallet.
Also, the current weight.rs file is already benched. Although,
solely on a probably non-representative machine.

Benches are currently only added to Altair. This should be
changed in the future.

* Integrate change requests, added usage of variable size in bench

* Clean up for linter