New Proof Integration
This commit implements the new proof into the pallet.
Hereby we use a costum base 2 merkle tree in order for us to allow
the generation of an arbitrary amount of contribution per
contributor. This is needed for the case that we want to reward
users on a different base than just "has contributed to crowdloan".