Integrates crowdloan pallets (claim and reward) (#456)
* New Proof Integration

This commit implements the new proof into the pallet.
Hereby we use a costum base 2 merkle tree in order for us to allow
the generation of an arbitrary amount of contribution per
contributor. This is needed for the case that we want to reward
users on a different base than just "has contributed to crowdloan".

* Move vesting to local deps

* Integrate crowdloan pallets (claim, reward) into dev rt

* Crowdloan reward benchmarks.

Benches run and weights.rs can be generated.
Wasm of RT builds.

NOTE: The developemnt rt contains also the benches for the
crowdloan-claim, which will not work and are hence commented out.

* WIP: Benches 'should' work but build fails

With build fails the build of the wasm with feature flag
'runtime-benchmarks' fails. The normal build does not fail...

Currently the benches are commented out in the development lib and
cargo file.

* Integrates own proofs lib

* WIP benches still not compiling, but std usage removed

* Compile dev rt again after rebase

* Claim benches NOT working. Signing needs std...

* Benches claim work now.

* Finall clean-up, remove conversion rate

* New benches from cloud machine

This benches represent the benchmarks made on a machine which
we run our validators with.

* Removed no-std-bin used for testing for no-std

* Added sp_std::Vec to inner