Redesign storage
Since we really just want an allowlist, we don't need to have a Status
type for each collator_id entry, and thus having the map rather as
HashSet makes more sense. The key has () as it's associated value type
and checking whether a collator is allowlisted, we just need to check
for the presence of said collator's id.