Fix onboard-chain error
We are passing the whole wasm file to the sudo tx call as an argument
and it fails because it is so long. Instead, we pass the path to the
wasm file.

Note: I am not sure this is actually supported since my parachain is not
producing blocks but I will investigate.