Restructure Permissions trait and implement TrancheInvestor differently
Now the permissions trait allows to check who called the
add and remove function and will check them with a statically defined
set of accounts that are allowed to to it for a given role.

The TrancheInvestor is now a vector and takes into account,
at which point in time he has been created.