Rt4 (#531)
* add uniques pallet to dev runtime

* add treasury pallet to dev and altair

* remove bounties

* swap pallet indexes on dev

* use AIR instead of CFG

* Added collator selection + enable governance calls + tx fees

* Fix match of call-filter

* Fix missing storage on genesis for session from collator selection

* First structure of rt upgrade for collator selection

* Compiles but no invulnerables passed

This is not tested. But the current migration compiles. But it would
leave the chain in a bad state, as no authorities are passed into
storage.

* Firs draft, can inject accounts

* Bump rt version to 1007

* feat: Add pallet-collator-allowlist (#535)

This pallet provides two extrinsics, one that allows sudo to
add a collator id to an allowlist, and another one that allows
sudo to remove them.

We have this pallet implementing `ValidatorRegistration`, which,
in addition to the default `Session` pallet implementation, also
checks for the presence of a collator id in this allowlist.

We do that to have tighter control over which collators get selected
per time windows, to avoid it defaulting to a FCFS setup until we
have chosen the right staking mechanism.

* Quick broom 🧹  (#541)

* Fix compilation error in migration/benchmarking

It fails with the following error:
expected struct `VestingInfo`, found struct `frame_support::BoundedVec`

* chore: Remove unused imports and drop dead code

* collator_allowlist: Benchmark + include it in the altair & dev runtimes (#542)

* benchmark: steps=50, repeat=100 (#543)

* chore: Group optional benchmark deps (#544)

To follow Ved's suggestion already applied to the collator-allowlist
pallet.

* Added invulnerables for our current collators

* fmt

* load validators from session

* revert period

* fmt

Co-authored-by: vedhavyas <vedhavyas.singareddi@gmail.com>
Co-authored-by: Miguel Hervas <miguel.hervas.lazaro@gmail.com>
Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>