[docker] Explicitly use compose file version 3.2
Newer versions of Docker interpret version `3` as `3.0`. This
causes loading of the file to fail, since we use features from
newer releases of the compose format. Explicitly setting it to
`3.2` allows the file to load.

`3.2` was chosen as it does not unnecessarily bump the minimum
Docker version required to run the compose file.