Use canonical para id where sensible
There is a trade-off here: if we change the para-id where we run our
development runtime, this won't work. On the other hand, we always use
2000 for that purpose, both in our local environment and in the xcm e2e
e2e tests. The pros is that we make it clear that the expected locations
are canonical, which is what other parachains will need to do to handle
our tokens or for us to handle others'.