Fix missing ability to close epoch with full-solution
Previously we would error out simply if the inspect solution
would error out with a full-solution.
This could lead to the inability of a pool to close an epoch,
depending on the amount of redeemptions and investements.

The change simply does not return early when the inspection of
the solution return an error, but rather turns the pool into
submission period.