Take a deposit when creating a pool
Previously, pool creation was free. This allows for spamming storage
with a large number of pools. To prevent this, we now take a deposit
when creating a pool.

To allow for the pool pallet itself to not care whether it is created
by a signed origin or a special origin (such as democracy), we first
attempt to take the deposit from the origin if possible. If the origin
cannot provide a deposit, we instead take it from the specified admin
account.