Cap number of roles per location in permissions pallet
Previously this was unbounded, which is a potential DOS vector against
the chain.

Rather than taking a deposit per-assigned-role, we instead set a
"sufficiently large" limit on the roles per location and expect that
the creation of that location will handle taking a deposit for the
entire thing.

In our current usage, a "location" is a pool, so this deposit needs to
be taken by the pool pallet.

This also refactors the permissions pallet slightly to avoid using a
`map_or` with a complex default. Instead we use the built-in substrate
`try_mutate` and take a more procedural code approach to the
permissions manipulation.