Fix/pool refactor to refactoring (#688)
* WIP: Does NOT compile. Draft refactor tranches

* Move tranches into new-type

THis commit moves the two tranche vectors of epoch tranches
and normal tranches into two respectuve new-types.

The newtypes implement a similar interface. Maybe in the future
we can abstract a trait-interface from there.

Fixes pool tests

Note: Also removes role aproval extrinsics.

* PR comments

* Rebase fixes, remove all as u32 in favor of tryinto

* restructure collect, tests

* Error out instead of implicit using last_epoch_executed

This commit now errors out in the pools::collect method instead
of implicitly using the last expoch that was executed as a maximum.

Closes #618

* Rename inner withdraw and deposit

* Add comnbinator methods, wrap reserve

Way to big. But introduces
* methods for tranches to combine sutff with them
* calculate prices
* accure debt
* reserve as its own struct in the pool

Not all tranche related functions are removed from pools.

* Fixing tests

Mostly concerned with wrong defaults and missing to reverse
tranches.

* Refactor combine methods of tranche for better generic usage

* EpochExecutionTranche wrap reversing and residual/non-residual

* Tranche wrap reversing and residual/non-residual

* rebalance tranches, ontroduce tranchetype

* move rebalance tranches to tranche

* Move calculate_risk_buffers to tranche.rs

* Remove impl of pow

* Change is_valid_tranche_change and input

* Fix tests

* Move solutions stuff to solutions.rs

* Fix tests and map underflow to Insufficient Curriency

* Make loons test compile

* Fix loan test

* Remove breaking tests change

* PR Requests

- remove impl on struct definition
- remove unused now parameter

* Remove not needed methods

* Rename methods tranche for residual/non-residual

* Fix and improve tranche price calculation

- use combine method instead of map and collect
- fix missing returning left-over assets
  -> Before if a tranche-value would be above the remaining assets, we would
     set the remaining assets to zero and return those. Which was wrong.

* use combine in order_as_currency

* Better error resolution in rebalance tranche

* Better error resolution in do_execute_epoch()

* Fix missing ability to close epoch with full-solution

Previously we would error out simply if the inspect solution
would error out with a full-solution.
This could lead to the inability of a pool to close an epoch,
depending on the amount of redeemptions and investements.

The change simply does not return early when the inspection of
the solution return an error, but rather turns the pool into
submission period.

* Expect instead of return, rename to residual_top_slice

* Fix pool tests

* address: #599

* Add comment to weight calculation

* Renaming methods as PR review suggested