Take a deposit when creating a pool (#685)
* Take a deposit when creating a pool

Previously, pool creation was free. This allows for spamming storage
with a large number of pools. To prevent this, we now take a deposit
when creating a pool.

To allow for the pool pallet itself to not care whether it is created
by a signed origin or a special origin (such as democracy), we first
attempt to take the deposit from the origin if possible. If the origin
cannot provide a deposit, we instead take it from the specified admin
account.

* Track the deposit taken for each pool

This allows us to change the deposit amount in the future and still
refund the correct amount for each pool.

* Store pool deposit information separately

* Add test and fix benchmarks