Fix underflow updating NAV when old PV was not computed correctly
Previously, it was possible to underflow the NAV if the old PV was not
computed correctly. This "should never happen", of course.

We now clamp the value to zero if it would have underflowed.

Any incorrect old PV will still result in a potentially miscomputed
intermediate NAV until the full NAV calculation is re-done.

Fixes #737