Additional RPC-Methods and Runtime-Apis (#796)
* Restructure and runtime api definitions

* Apis and rpc impls

* Fix runtime impls

* Finish loans-api, extend dev node with it

* Empty impl for pools-api, extend dev node with it

* Impl pools-api in dev runtime

* Fix missing debug in tests

* Use TrancheIds instead of TrancheLocs in public RPC APIs

Consumers of the API should always use an Id, never an index. Indexes
may change underneat an API consumer at any block if the tranches
change, while Ids are stable.

This change enforces that correctness, and simplifies client code
which no longer needs to wrap Ids in Locs

* Remove pool_value RPC method

This is panicking for unknown reasons and can easily be computed on
the client instead. We're giving up on it for now ¯\_(ツ)_/¯

* Address feedback

* Remove unused `pool_value` from Runtime API
* Remove `pool_` prefix, as it is redundant at the RPC level

* Merge remote-tracking branch 'upstream/parachain' into feature/rpc-extensions

* Remove loans RPC methods

These aren't totally working and aren't needed for the first release

* Update src/rpc/pools.rs

Co-authored-by: Jeroen Offerijns <jeroen@offerijns.com>
Co-authored-by: Branan Riley <branan@centrifuge.io>
Co-authored-by: Jeroen Offerijns <Offerijns@users.noreply.github.com>