Unify fixed point & Interest Accrual (#789)
* feat: track repaid amount per loan

* chore: format

* fix: use correct repayment amount

* chore: add test for repaying more than debt

* feat: set up interest rates pallet

* feat: start on debt calculation methods

* feat: work on adding to runtime

* feat: inject interest accrual in loans pallet

* feat: implement debt adjustment

* feat: work on updating pv calculations using debt

* fix: type

* chore: format

* fix: rename debt method, fix some bugs

* chore: format

* fix: move adjustment to common-types

* fix: squash some bugs

* docs: add interest accrual explainer comment

* fix: update accumulated rate on retrieving current debt

* chore: format

* fix: typo

* Convert Loans pallet from Amount to Balance

This is mostly a straightforward conversion. Some math had to be
slightly restructured due to now using `FixedPoint*Integer`
multiplies instead of `FixedPoint*FixedPoint`.

Some tests still fail - they'll be fixed up in an upcoming commit.

* Fix bugs, tests and warnings

* Fix benchmarks

* Fix loans benchmarks for pool deposit

* Use NormalizedDebt as an associated type of the interest pallet

This removes the need to specify a separate Normalized type for the
loan and interest pallets. Using an associated type in this way forces
us to treat the type as opaque.

* Make loans tests compile again

Still broken in several places due to assumptions the tests make
about interest rate and principal debt, compared to how the
accrual pallet now normalizes rates.

* (tmp) Allow fetching the previous value for an interest rate

This is a hack to get the loans pallet back to baseline
functionality. It needs to be replace with some sort of caching
system (trading space for time), otherwise we lose the benefits
of the interest rate pallet.

* Don't try to run the interest-accrual graph as a test

* Use current debt when computing max borrow/repay amounts

* Test fixes for loans refactors

This gets all the tests finally passing for the loans refactor
to use the interest accrual pallet.

* address feedback

* Update pallets/loans/src/math.rs

Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>

* Add comment for adjustment enum

* Undo type change

* Fixups from bad rebase

Co-authored-by: Jeroen Offerijns <jeroen@offerijns.com>
Co-authored-by: Jeroen Offerijns <Offerijns@users.noreply.github.com>
Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>