Store the block number at which a loan was closed
Currently we are keeping details of all closed loans available in
storage. This is eventually going to get Too Big, and we'll need to
migrate to looking at historical data in some way.

Keeping the block number at which a loan was closed will allow future
client code to query the old storage for the loan details, rather than
needing to keep the ClosedLoans storage around indefinitely.