Move interest accrual to `on_initialize` (#870)
* Accrue interest rates in on_initialize

This nets us a 60% perf improvement in nav updates over the
current implementation (at the cost of the on_initialize weight)

This also modifies some logic around penalty interest rates
temporarily until that logic can be made more correct.

Still do to:
* Benchmark and test the migration
* Benchmark the on_initialize
* Fix how we calculate penalty rates and re-enable them

* Re-introduce write-off penalties

This is messy and needs to be cleaned up. Our logical nesting is
way too deep in several functions now.

* Add interest-accrual migration file

* Remove panics in interest-accrual migration/on_initialize

* Benchmark rate accrual for on_initialize and migration

* Fix altair runtime compilation

* Add testing for interest-accrual migration

* Migrate interest_accrual and loans in lockstep

The `loans` pallet needs to reference all the rates that it uses
on the runtime upgrade. This removes the `on_runtime_upgrade` of
interest accrual, and instead uses a coordinated upgrade
implemented at the runtime level.

* Remove unused rates during migration

* Clean up calculation of rate+penalty

* Remove unnecessary nested storage access closures

* Update nix Cargo sha

* Minor fixes & Code review addressed

* Update nix cargosha256

Co-authored-by: Jeroen Offerijns <Offerijns@users.noreply.github.com>