Make compile again with asterix
the following approach would work, but it does not as
currently parity does not take over all trait bounds on all
implementations on Pallet<T>.

So I wanna have this state stored and the next commit will revert
it to a state that does actually compile.

Once, https://github.com/paritytech/substrate/issues/12053 is
resolved, we can simply revert the next commit and should be fine.