Enable clippy in CI and Fix or Allow all clippy warnings (#982)
* Enable clippy in CI

* Fix clippy::assign-op-pattern

* Fix clippy::bool-comparison

* Fix clippy::into-iter-on-ref

* Fix clippy::int-plus-one

* Fix clippy::len-without-is-empty

* Fix clippy::manual-map

I went a bit further with this one, since there were some obvious
cleanups:

* What clippy called a manual implementation of `Option::map` was
  really an implementation of `Option::cloned`.
* The outer `if let` was just `Option::and_then`, so I replaced that
  as well

* Fix clippy::map-flatten

* Fix clippy::match-single-binding

* Fix clippy::redundant-field-names

* Fix clippy::redundant-static-lifetimes

* Fix clippy::search-is-some

* Fix clippy::tabs-in-doc-comments

* Allow clippy::too-many-arguments

* Fix clippy::unused-unit

* Fix clippy::vec-init-then-push

* Fix clippy::borrowed-box

* Fix clippy::useless-vec

* Fix clippy::bind-instead-of-map

* Fix clippy::comparison-chain

* Fix clippy::iter-cloned-collect

* Fix clippy::ptr-arg

* Fix clippy::map-clone

* Allow clippy::result-unit-err

This error type is ultimately fed up through a trait layer which is
generic over the Error type. The code in this file would become
unnecessarily complex if we had to convert between errors somehow.

* Fix clippy::len-zero

* Fix Tranche add/remove logic

Found while fixing clippy::len-zero issues

Add issues:
* Boundary in `ensure` was reversed
* Unnecessary guarding against inserting at idx `0` of an empty array
* Unnecessary guarding against inserting at `idx==len` of a non-empty array
* Once boundary checks were fixed, branches were deduplicated and simplified

Remove issues:
* Boundary in `ensure` was reversed
* Removing idx 0 from an empty tranche set was silently accepted, instead of erroring
* Once boundary checks were fixed, branches were deduplicated and simplified

* Fix clippy::match-like-matches-macro

* Allow clippy::type-complexity

Cleaning this up would be a larger refactor than is warranted for
initial clippy cleanups. Allow for now, and we can clean up as-needed

* Fix clippy::useless-conversion

* Fix clippy::needless-return

* Fix clippy::clone-on-copy

* Allow clippy::unit-arg

This issue only appears in libs/types/src/permissions.rs

However, it appears _a lot_ in that file. Leaving it be for now.

* Fix clippy::needless-borrow

* Allow clippy::identity-op

We use this for the `1 * CFG` or `1 * HOURS` pattern in our code. It's
sort of a false-positive in those cases.

* Fix clippy::redundant-clone

* Fix/Allow clippy::or-fun-call

This fixes most instances of this lint. However, there are 22 of them
in the pools pallet. Leaving those for another time.

* Allow all clippy lints on pallet-anchors

Per @mikiquantum, this pallet is getting a major refactor soon. No
sense messing with the current code.

* Remove unused genesis structures

This was found by `clippy::derivable-impls` due to the manual
implementations of `Default` for these genesis structs. Rather than
just clean that up, they've been totally removed as they are unused.