Fix Bug in Pools::deposit() logic
The bug was that we did assume, that the maximum amount of money from
a deposit a junior tranche is entitled to is its debt. BUT
this is not true. It's always the leftover of what other tranches
where not entitled to.

This is not a bug, as during rebalancing we do NOT rely on
the debt calculations of the junior tranche. Anyways, this keeps
the system longer in consistent behaviour.