Update to Polkadot v0.9.29 (#1041)
* Update to polkadot v0.9.29 (#1040)

wip

* pallet-pools: Enable default-features in dev-deps

* Align all serde depds to v1.0.119

* cargo: update to fetch latest internal revisions

* Set all dev.dependencies' default-features=true

* try: getrandom = { version = "0.2", features = ["js"] }

Let's commit this and see if the CI gets the same build error.

* sort std

* try: Add getrandom with default-features = false

This moves the issue down to failed to run custom build command for
`secp256k1-sys v0.6.1`. This GH issue suggests it may be an issue with
M1 support: https://github.com/tomusdrw/rust-web3/issues/543

* Fix wrong 'default-features = true'

* Sort benchmarks

* Sorting

* Fix weights in dev runtime

* Fix command.rs

* flake.nix

* Remove std entries left behind

* Clean up explicit dependency on getrandom

Having it on the root Cargo.toml is enough :)

* fmt

* Adjusting mocks and runtimes

* Adjust to new execute_block function and replace Weights in mocks

* flake.nix

* Trying to adjust to new try_execute_block

* Adding try-runtime to std

* Drop getrandom 💡

I just realized that we probably only needed this because we were
importing dependencies from pallet-claims with the default-features
enabled, which caused getrandom from being pulled with enabled std WHILE
trying to build it for wasm, which caused the failing compilation
error. Now that those dependencies are imported correctly with
default-features = false, we in fact don't need this getrandom
workaround anymore :)

* Update Cargo.lock

* flake.nix

* Remove logs

* Adding try-runtime

* Adjust try-runtimes

* Add try-runtimes

* Use latest chainbridge

* Update integration tests

* flake.nix

* Adjusting interest-accrual to new Weight type

* Fix typo

* Small fixes

* loans benchmarks: ensure generated rates match our rules
* integration tests: Do not create duplicate NFT class

I have no idea how either of thesee was working before ¯\_(ツ)_/¯

* set dev-defaults to true, update fudge timers

* cargo fmt

* Update weight scripts

* Remove

Co-authored-by: Bastian Gruber <foreach@me.com>