Use a vector in interest-accrual
This converts the interest-accrual pallet to use a vector for storage
of interest rates, immediately giving us less block time spent in
`on_initialize` for the accrual step.

This also includes the storage migration, and cleans up some stale
migration code in the interest-accrual pallet.

This is the first part of #1021. There is also work to be done in
pallet-loans to take advantage of this change.