Convert interest-accrual storage to a vector (#1157)
* Use a vector in interest-accrual

This converts the interest-accrual pallet to use a vector for storage
of interest rates, immediately giving us less block time spent in
`on_initialize` for the accrual step.

This also includes the storage migration, and cleans up some stale
migration code in the interest-accrual pallet.

This is the first part of #1021. There is also work to be done in
pallet-loans to take advantage of this change.

* Use the interest rate vec when calculating NAVs

* Update loans benchmarks for vector rate storage

This updates the loans benches to account for the new computational
complexity of the linear search for accrued interest rates.

The weights still need to be re-generated, and the interest accrual
APIs need to be updated to return the actual number of items.

* Use updated benchmarks for loans pallet

This updates the loans pallet to the new benchmark interface. The
interest-accrual pallet still does not expose the number of rates,
so the final weight of all loans extrinsics uses the max.

* Fix taplo and nix

* Storage cleanups

* Use a common accessor method to avoid repeated iterator logic
* Use try_mutate where it makes sense

* Account for storage writes in drain()

* Add try-runtime test for migration

* Storage is now initialized at V2 for new genesis blocks

* Bump flake.nix

* DRY up some loans pallet init code in benchmarks

* Address minor feedbacks