Fix loans benchmarks in debug builds
`pallet-system` assumes a block number of at least 1, and panics
with a subtraction underflow when we use block number 0. This
doesn't affect the behavior of the benchmarks, since the measured
code is not affected by the block number. It does prevent running
the benches in debug builds, since the underflowing subtraction
panics when built in debug mode.