Evm part 2 (#1349)
* Cleanup un-addressed feedback from initial EVM PR

* Split some EVM parts out to common

* Add EVM to Altair and Centrifuge runtime

This adds EVM support to the Altair and Centrifuge runtimes,
with the following caveats:

* All EVM pallet functionality is filtered on the substrate side
* The EVM `create` RPC is filtered in the runtimes

This allows EVM RPC calls to be made to invoke precompiles or
smart contracts deployed by Governance (root).

* Add EVM node support to altair & centrifuge

* Remove non-evm service fns

* Remove unused dynamic_fee data provider

* fix lints

* Subalfred

* Filter EVM create at transaction check time

The `create` runtime API is used for the `call` and `estimate_gas`
RPCs, but not for transactions stored in a block.

Transactions in a block are encoded as wrapped extrinsics. This
wrapping is done by the `fp_rpc::ConvertTransaction` runtime API. I
don't believe filtering at this conversion point is safe, since a
malicious node could choose to create a transaction differently and
still add it to the pool.

Instead, we do the filtering on conversion from an "unchecked" to a
"checked" extrinsic. For both Substrate and Ethereum transactions,
this conversion typically involves just validating that transaction
fees can be paid, that signatures are valid, etc. We add additional
logic to catch EVM create invocations and indicate that such
transactions are invalid.

* Update runtime/altair/src/evm.rs

Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>

* Add EVM configuration to CLI

* Move WEIGHT_PER_GAS constant to common

* Add EVM Chain ID migration to Centrifuge and Altair

* Allow clippy lint in code from Frontier template

* Unify account conversion for EVM and Connectors

This brings EVM account conversion and connectors DomainAccount
conversion into one single location.

Moving forward, this can be expanded to support account conversions
across multiple sources using xcm v3 MultiLocations. But as we are
still on XCM v2 we don't have the necessary XCM primitives yet.

* Add a few more comments

* Add tests for account conversion

* Update endowed EVM accounts to account for new conversion

---------

Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>